requirements.txt
GitPython==3.1.18
docker-compose==1.29.2
kubernetes==17.17.0